fix percentage bug & fix exponentially high wants values

+++ src-tauri/src/config.rs
@@ -8,12 +8,12 @@ const CONFIG_PATH: &str = "./game_config.toml";
 pub struct Config {
     pub starting_population: i32,
 
-    pub no_education_chance: f32,
-    pub high_school_diploma_chance: f32,
-    pub college_chance: f32,
-    pub associate_degree_chance: f32,
-    pub bachelors_chance: f32,
-    pub advanced_degree_chance: f32,
+    pub no_education_chance: i32,
+    pub high_school_diploma_chance: i32,
+    pub college_chance: i32,
+    pub associate_degree_chance: i32,
+    pub bachelors_chance: i32,
+    pub advanced_degree_chance: i32,
 }
 
 pub fn load_config() -> Config {
